backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ma-mna-analyzer"
version = "0.1.0"
description = "FastAPI backend to automate public-document M&A opportunity analysis for two European listed companies via the OpenAI API."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "lxml>=5.3.0",
  "openai>=1.0.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "uvicorn[standard]>=0.30.6"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
